requirements.txt
absl-py==2.3.1
amqp==5.3.1
asgiref==3.11.0
astunparse==1.6.3
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
exceptiongroup==1.3.1
ExifRead==3.5.1
filelock==3.20.1
flatbuffers==25.9.23
fsspec==2025.12.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
kombu==5.6.1
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psycopg2==2.9.11
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
sympy==1.14.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.1
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
